requirements.txt
Flask==2.1.0
requests==2.27.1
beautifulsoup4==4.9.3
gunicorn==20.1.0
python-dotenv==1.0.0
gevent-websocket==0.10.1
websockets==12.0